Wrote code for redirect, appears to work but unsure
